pyproject.toml
[project]
name = "whaleradar"
version = "0.1.0"
description = "CryptoWhaleTracker — Real-time whale movement detection and alert system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Data collection
    "websockets>=14.0",
    "httpx>=0.28",

    # Kafka
    "confluent-kafka>=2.6",

    # Spark
    "pyspark>=3.5",

    # AI / LLM agents
    "langgraph>=0.2",
    "langchain-openai>=0.3",
    "langchain-core>=0.3",
    "openai>=1.60",

    # Telegram
    "python-telegram-bot>=21.0",

    # Dashboard
    "streamlit>=1.40",

    # Config & validation
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",

    # HDFS
    "hdfs>=2.7",

    # Utilities
    "orjson>=3.10",
    "structlog>=24.0",
]

[project.scripts]
whaleradar = "whaleradar.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whaleradar"]
